Add support for user-supplied configuration.
This change adds three configurable options to the script that can be
customized by the user to control the search thresholds for creating,
removing bookmarks, and optionally pausing after jumping to a bookmark.

Also fixes a latent bug with the time threshold for seeking bookmarks.
Seeking forward wasn't using the threshold so it could find the current
bookmark and be stuck. This surfaces prominently when pausing on jump.